login and registration end to end tests
